feat: update header to 1.21.111

+++ src/mc/world/level/block/FlameOdds.h
@@ -2,7 +2,7 @@
 
 #include "mc/_HeaderOutputPredefine.h"
 
-enum class FlameOdds : int {
+enum class FlameOdds : ushort {
     Instant = 60,
     Easy    = 30,
     Medium  = 15,
